add boss.py inventory.py player.py
